+++ parser/parser.h
@@ -1,0 +1,18 @@
+//
+//  parser.h
+//  langtons_ant
+//
+//  Created by Srinivas Dhanwada on 8/28/16.
+//  Copyright © 2016 dhanwada. All rights reserved.
+//
+
+#ifndef parser_h
+#define parser_h
+
+#include <stdio.h>
+
+extern int BORDER_MODE;
+
+void parse_args(int argc, char** argv);
+
+#endif /* parser_h */